// backend/src/instruments/enums/instrument.enum.ts
export enum InstrumentCategory {
  STRING = 'string',
  WIND = 'wind',
  PERCUSSION = 'percussion',
  KEYBOARD = 'keyboard',
  OTHER = 'other',
}

export enum InstrumentStatus {
  AVAILABLE = 'available',
  IN_REPAIR = 'in_repair',
  SEALED = 'sealed',
  IN_PERFORMANCE = 'in_performance',
  BORROWED = 'borrowed',
}